Fix the undefined symbol in global_config.h
Currently, applications using GetDevicePluginPath(), defined in
global_config.h, will fail to build if "-DLIBDIR_PREFIX" is not one of
the build flags.

To fix the issue, use the cmake configure_file() directive for
global_config.h.in (generated as global_config.h during the build), and
the appropriate install dir prefix will be substituted for
"@LIB_INSTALL_DIR@".

Change-Id: I9aca9bf222ba3ce669d29903ac37cedbdc09a070
Signed-off-by: Patrick McCarty <patrick.mccarty@linux.intel.com>